Fix the problem with getting both a wx.EVT_LIST_ITEM_DESELECTED and wx.EVT_LIST_ITEM_SELECTED on a selection change.
By using the fact that the event queue is FIFO, we just tell wxPython to call
the DESELECTED handler at a later date (using wx.CallAfter). Then in the
handler we check if the selection has actually changed - this will only be the
case when there was no associated SELECTED event.